backend/parsers/parcel_automat/requirements.txt
fastapi
uvicorn
pydantic
pydantic-settings
httpx
pyserial
python-dotenv
opencv-python-headless
pyzbar
numpy
pytest
pytest-asyncio
pytest-cov
pytest-mock